+++ CHANGELOG.rst
@@ -9,6 +9,8 @@ Version 0.4.0
 
 Minor release, unreleased
 
+- Add ability to set attributes on ``PoolConnection``.
+
 Version 0.3.0
 -------------